// src/webparts/projectDetails/components/Dialogs/ShareDocumentDialog.module.scss.d.ts
export interface IDialogModuleScss {
  dialog: string;
  dialogContent: string;
  field: string;
  uploadSection: string;
  uploadButton: string;
  userSection: string;
  label: string;
  userList: string;
  userItem: string;
  avatar: string;
  userName: string;
  checkbox: string;
  note: string;
  infoBox: string;
  infoIcon: string;
  infoText: string;
  errorMessage: string;
  errorIcon: string;
  primaryButton: string;
  tableContainer: string;
  table: string;
  checkboxColumn: string;
  nameColumn: string;
  locationColumn: string;
  row: string;
  nameCell: string;
  fileIcon: string;
  emptyState: string;
  emptyIcon: string;
  selectedInfo: string;
  loadingContainer: string;
}

declare const styles: IDialogModuleScss;
export default styles;